third_party/WebKit/LayoutTests/animations/animation-none-keyword.html

<!doctype html>
<html lang="en">
<head>
  <title>Testing animation on a none keyword</title>
  <style>
    #test {
      max-width: none;
      -webkit-animation-duration: 2s;
      -webkit-animation-timing-function: linear;
      -webkit-animation-name: none-test;
    }

    @-webkit-keyframes none-test {
      25% { max-width: 10px; }
      75% { max-width: 10px; }
    }
  </style>
  <script src="resources/animation-test-helpers.js"></script>
  <script>
    const expectedValues = [
      // [animation-name, time, element-id, property, expected-value, tolerance]
      ["none-test", 0, "test", "max-width", "none", 0],
      ["none-test", 0.25, "test", "max-width", 10, 0],
      ["none-test", 1, "test", "max-width", 10, 0],
      ["none-test", 1.75, "test", "max-width", "none", 0],
      ["none-test", 2, "test", "max-width", "none", 0],
    ];

    runAnimationTest(expectedValues);
  </script>
</head>
<body>
<div id="test">Animating max-width from/to none.</div>
</div>
</body>
</html>
