third_party/WebKit/LayoutTests/platform/win-xp/virtual/web-animations-css/animations/animation-none-keyword-expected.txt

FAIL - "max-width" property for "test" element at 0s expected: 10 but saw: none
PASS - "max-width" property for "test" element at 0.25s saw something close to: 10
PASS - "max-width" property for "test" element at 1s saw something close to: 10
PASS - "max-width" property for "test" element at 1.75s saw something close to: none
PASS - "max-width" property for "test" element at 2s saw something close to: none

Log:
20.999999999958163 - runAnimationTest
23.999999999887223 - First webkitAnimationStart event fired
24.999999999863576 - Finished waiting for animations to start
24.999999999863576 - Pausing at time: 0, active animations: 1
24.999999999863576 - Checking expectation: ["none-test",0,"test","max-width",10,0]
28.00000000002001 - Pausing at time: 250, active animations: 1
28.00000000002001 - Checking expectation: ["none-test",0.25,"test","max-width",10,0]
28.00000000002001 - Pausing at time: 1000, active animations: 1
28.00000000002001 - Checking expectation: ["none-test",1,"test","max-width",10,0]
28.999999999996362 - Pausing at time: 1750, active animations: 1
28.999999999996362 - Checking expectation: ["none-test",1.75,"test","max-width","none",0]
28.999999999996362 - Pausing at time: 2000, active animations: 1
28.999999999996362 - Checking expectation: ["none-test",2,"test","max-width","none",0]
28.999999999996362 - Ending test
